Add files via upload
The malloc is also known as the memory allocation function. malloc() dynamically allocates a large block of memory with a specific size. It returns a void type pointer and is cast into any form.
The calloc() function allocates a specific amount of memory and initializes it to zero. The function can be cast to the desired type when it returns to a void pointer to the memory location.